Installation and Removal Checks.
Checks for the existence of files and provides informational feedback while the script is running

+++ deploy_docker.sh
@@ -27,7 +27,7 @@ docker_packages=(
   'docker.io'
   'docker-doc'
   'docker-compose'
-  'docker-comose-v2'
+  'docker-compose-v2'
   'podman-docker'
   'containerd'
   'runc'
@@ -43,27 +43,52 @@ usage() {
 }
 
 remove() {
+  echo -e "${INFO}Removing docker binaries, gpg key, and apt list."
   # Uninstall docker engine binaries
   for pkg in "${docker_packages[@]}"; do
-    sudo apt remove $pkg
+    sudo apt remove $pkg -y
   done
+
+  if [ -e /etc/apt/sources.list.d/docker.list ]; then
+    sudo rm -v /etc/apt/sources.list.d/docker.list
+  else
+    echo "${INFO}Docker apt list not found.. skipping"
+  fi
+
+  if [ -e /etc/apt/keyrings/docker.asc ]; then
+    sudo rm -v /etc/apt/keyrings/docker.asc
+  else
+    echo "${INFO}Docker gpg key not found.. skipping"
+  fi
+
+  echo -e "${INFO}Successfully cleaned up."
+  exit 0
 }
 
 install() {
   sudo apt update
-  sudo apt install ca-certificates curl
+  sudo apt install ca-certificates curl -y
   sudo install -m 0755 -d /etc/apt/keyrings
   sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
-  sudo chmod a+r /etc/apt/keyrings/docker.asc
+  if [ -e /etc/apt/keyrings/docker.asc ]; then
+    sudo chmod a+r /etc/apt/keyrings/docker.asc
 
-  echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc]\
-    https://download.docker.com/linux/ubuntu $(. /etc/os-release && echo "$VERSION_CODENAME") stable" |
-    sudo tee /etc/apt/sources.list.d/docker.list >/dev/null
+    echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc]\
+      https://download.docker.com/linux/ubuntu $(. /etc/os-release && echo "$VERSION_CODENAME") stable" |
+      sudo tee /etc/apt/sources.list.d/docker.list >/dev/null
 
-  sudo apt update
-  sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+    sudo apt update
+    sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin -y
+  else
+    echo "${WARN}Docker gpg key not found.. Exiting"
+    exit 1
+  fi
+
+  echo -e "${SUCCESS}Docker successfully installed. Running hello-world.."
 
   sudo docker run hello-world
+
+  exit 0
 }
 
 while getopts 'hir' flag; do
